Clean up and refactor code
Remove unused methods and refactor export command handler into separate class
