+++ src/main/java/seedu/flowcli/exception/FlowCLIExceptions.java
@@ -54,10 +54,4 @@ public class FlowCLIExceptions {
             super("Failed to write to file: " + message);
         }
     }
-
-    public static class NoViewToExportException extends Exception {
-        public NoViewToExportException() {
-            super("No view to export. Please filter, sort, or list tasks first, or specify export parameters.");
-        }
-    }
 }
